+++ public-api/CHANGELOG.md
@@ -1,5 +1,8 @@
 # `public-api` changelog
 
+## v0.43.0
+* Support `nightly-2025-01-25` and later.
+
 ## v0.42.0
 * Render `_` argument names in function declarations
 * Add `PublicItem::id(&self) -> rustdoc_types::Id` and `PublicItem::parent_id(&self) -> core::option::Option<rustdoc_types::Id>`